requirements.txt
asgiref==3.8.1
base58==2.1.1
bleach==6.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.2.15
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-daraja==1.3.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-suit==0.2.28
django-summernote==0.8.20.0
djangorestframework==3.15.2
environ==1.0
gunicorn==23.0.0
idna==3.8
nanoid==2.0.0
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
